fixing up logic
this was made too complicated. I want to finish it but i dont want to release crap. So i changed the way it works. No more base price. now if a category has a base_item then it means something in that category is required. If it does not then you do not need something from that cateogry